Implementando tela de relatório de estoque de produtos da cozinha e aplicando herança à telas de relatórios de estoque (#21)
* Implementando histórico de produtos

* Implementado super classe e herança para relatórios de estoque de chope e produtos

* Implementado super classe e herança para histórico de chope e produtos

* Implementando tela de relatório de estoque de produtos da cozinha

* Consertando título das telas de relatório de estoque

* Alterando tipo do atributo 'stock' da entidade de produto de 'Long' para 'Double'

---------

+++ backend/src/main/java/br/com/chadschoperia/service/dto/ProductDto.java
@@ -27,12 +27,12 @@ public class ProductDto implements Serializable {
 	private Long restockThreshold;
 
 	@NotNull(message = "product.stock.not_null")
-	private Long stock;
+	private Double stock;
 
 	@NotNull(message = "product.barcode.not_null")
 	private Long barcode;
 
-	public ProductDto(Long id, String name, Long restockThreshold, Long stock, Long barcode) {
+	public ProductDto(Long id, String name, Long restockThreshold, Double stock, Long barcode) {
 		this.id = id;
 		this.name = name;
 		this.restockThreshold = restockThreshold;
@@ -40,7 +40,7 @@ public class ProductDto implements Serializable {
 		this.barcode = barcode;
 	}
 
-	public void setStock(Long stock) {
+	public void setStock(Double stock) {
 		if (stock < 0) {
 			throw new OutOfStockException("product.out_of_stock");
 		}
